Fix licenses (#246)
fixed license and author information, added license info to gmi, cleanup motion gmi

+++ src/gomc/internal/task/stat.go
@@ -1,3 +1,5 @@
+// Copyright (C) 2026 Sascha Ittner <sascha.ittner@modusoft.de>
+// License: GPL Version 2
 package task
 
 import (
@@ -122,10 +124,10 @@ func (t *Task) BuildStat() *emcstat.StatFull {
 	stat.Motion.Velocity = ms.RequestedVel
 	stat.Motion.CurrentVel = ms.CurrentVel
 	stat.Motion.DistanceToGo = ms.DistanceToGo
-	stat.Motion.MotionId = ms.Id
-	stat.Motion.MotionLine = ms.Id
+	stat.Motion.MotionId = int32(ms.Id)
+	stat.Motion.MotionLine = int32(ms.Id)
 	stat.Motion.MotionType = ms.MotionType
-	stat.Task.MotionLine = ms.Id
+	stat.Task.MotionLine = int32(ms.Id)
 	stat.Motion.Dtg = emcstat.Position{
 		X: ms.Dtg.X, Y: ms.Dtg.Y, Z: ms.Dtg.Z,
 		A: ms.Dtg.A, B: ms.Dtg.B, C: ms.Dtg.C,
